minimap -> minimap_player/mimimap_ray
