+++ Makefile
@@ -57,12 +57,14 @@ SRC_FILES = $(SRC_DIR)/main.c \
 			$(PARSER_UTILS_DIR)/map_utils.c \
 			$(PARSER_UTILS_DIR)/string_utils.c \
 			\
+			$(RENDER_DIR)/cast_rays.c \
+			$(RENDER_DIR)/draw_primitives.c \
+			$(RENDER_DIR)/minimap_player.c \
+			$(RENDER_DIR)/minimap_ray.c \
 			$(RENDER_DIR)/minimap.c \
 			$(RENDER_DIR)/render_utils.c \
 			$(RENDER_DIR)/render.c \
-			$(RENDER_DIR)/draw_primitives.c \
 			$(RENDER_DIR)/setup.c \
-			$(RENDER_DIR)/cast_rays.c \
 
 
 OBJ_FILES = $(SRC_FILES:$(SRC_DIR)/%.c=$(OBJ_DIR)/%.o)
